arreglando fk en migraciones de finanzas

+++ AgrosoftLaravel/database/migrations/2025_06_26_000020_create_desechos_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('desechos', function (Blueprint $table) {
             $table->id();
-            $table->integer('fk_Cultivos');
+            $table->unsignedBigInteger('fk_Cultivos');
             $table->unsignedBigInteger('fk_TiposDesecho');
             $table->string('nombre');
             $table->text('descripcion');
